Move log.js into utils folder
