+++ lib/code-manager.js
@@ -5,7 +5,7 @@ import { Point, Range } from 'atom';
 import escapeStringRegexp from 'escape-string-regexp';
 import _ from 'lodash';
 
-import log from './log';
+import log from './utils/log';
 
 export default class CodeManager {
   constructor() {
