+++ lib/inspector.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import * as transformime from 'transformime';
 
-import log from './log';
+import log from './utils/log';
 import store from './store';
 import reactFactory from './utils/react';
 import InspectorComponent from './components/inspector';
