+++ lib/kernel-manager.js
@@ -10,7 +10,7 @@ import Config from './config';
 import WSKernel from './ws-kernel';
 import ZMQKernel from './zmq-kernel';
 import KernelPicker from './kernel-picker';
-import log from './log';
+import log from './utils/log';
 import store from './store';
 import { grammarToLanguage } from './utils';
