+++ lib/kernel-picker.js
@@ -3,7 +3,7 @@
 import { SelectListView } from 'atom-space-pen-views';
 import _ from 'lodash';
 
-import log from './log';
+import log from './utils/log';
 
 // View to display a list of grammars to apply to the current editor.
 export default class SignalListView extends SelectListView {
