+++ lib/kernel.js
@@ -8,7 +8,7 @@ import store from './store';
 
 import WatchSidebar from './watch-sidebar';
 import HydrogenKernel from './plugin-api/hydrogen-kernel';
-import log from './log';
+import log from './utils/log';
 
 export default class Kernel {
   @observable executionState = 'loading'
