+++ lib/main.js
@@ -22,7 +22,7 @@ import Inspector from './inspector';
 import AutocompleteProvider from './autocomplete-provider';
 import HydrogenProvider from './plugin-api/hydrogen-provider';
 import reactFactory from './utils/react';
-import log from './log';
+import log from './utils/log';
 
 const Hydrogen = {
   config: Config.schema,
