+++ lib/result-view.js
@@ -3,7 +3,7 @@
 import { CompositeDisposable } from 'atom';
 import { createTransform } from 'transformime';
 
-import log from './log';
+import log from './utils/log';
 
 const transform = createTransform();
