+++ lib/signal-list-view.js
@@ -4,7 +4,7 @@ import { SelectListView } from 'atom-space-pen-views';
 import _ from 'lodash';
 
 import WSKernel from './ws-kernel';
-import log from './log';
+import log from './utils/log';
 import store from './store';
 
 // View to display a list of grammars to apply to the current editor.
