+++ lib/watch-view.js
@@ -3,7 +3,7 @@
 import { TextEditor } from 'atom';
 
 import ResultView from './result-view';
-import log from './log';
+import log from './utils/log';
 
 export default class WatchView {
