+++ lib/zmq-kernel.js
@@ -6,7 +6,7 @@ import v4 from 'uuid/v4';
 
 import Kernel from './kernel';
 import InputView from './input-view';
-import log from './log';
+import log from './utils/log';
 
 export default class ZMQKernel extends Kernel {
   constructor(kernelSpec, grammar, connection, connectionFile, kernelProcess) {
